Added more information to error messages printed in some parts of the code. Added a few new debugging log messages.
The PGX format appears to allow the plus/minus character in the header
to be omitted (which would then default to plus).  This was not properly
handled before and has now been corrected.

One of the PGX error messages incorrectly claimed to be for the BMP codec.
This has been fixed.

Started to add some infrastructure for performing decoder tests.
This will remain disabled until completed.