Sometimes the abs function was being used in places where a type other than int was involved (e.g., int_fast32_t). This could lead to problems, depending on the size of int relative to other integral types. (One non-so-important assertion in jiv.c was using abs with doubles.) These problems have been fixed by using JAS_ABS instead of abs.
The parameter JAS_IMAGE_INMEMTHRESH has been increased to 256 MB.
It was previously set to 16 MB, which is quite low for most systems today.

Some additional assertions were added in the asl/asr code to make some
attempt at ensuring that shifting of negative numbers (which is in the
land of UB) works as expected.

A few int types were changed to jpc_fix_t in the JPC codec.
This change is needed to avoid potential problems on systems where int
and jpc_fix_t are different.