Added a command line option to imginfo to allow for arbitrary decoder options to be specified.
Adjusted the debug level associated with some log messages.

Added an option to the JPEG decoder that makes it print the JPEG library
version to standard output and return without decoding
(i.e., return with an error).
Also, added a debug log message that includes the JPEG library version.

Fixed a possible use of uninitialized data in the JPEG decoder.

Modified the run_test_1 script to try to better accommodate different
versions of the JPEG library (which have different behaviors for some
of the decoding tests).