CMakeLists.txt: enable -fvisibility=hidden for internal symbols
Don't export internal symbols in the public symbol table.  This
reduces the number of exported symbols from 348 to 121.

The same is already being done on Windows by using
`__declspec(dllexport)`, and all functions not annotated with it will
be inaccessible by programs using JasPer.

Strangely, this refuses access to lots of functions in the public
headers.  I wonder why these were declared in the public headers.

To avoid breakages with programs using those internal symbols, this
feature is optional and must be enabled with the cmake option
`-DJAS_ENABLE_HIDDEN=true`.