jas_seq, jpc_fix: add option to use 32-bit integers
Those libraries use the C99 type `int_fast32_t`, which translates to
`int64_t` on AMD64.  This type obviously consumes twice as much memory
as `int32_t`, and these two libraries are mostly about allocating bulk
buffers containing these values.

More memory use for large buffers means a larger working set, which
results in reduced CPU cache efficiency.  So in these cases,
`int_fast32_t` is not actually the "fastest" integer, but the slowest.

This is a "perf stat imginfo -f balloon.jp2" with the 64 bit ABI (on
an Intel Xeon E5-2630):

       1010.691814      task-clock:u (msec)       #    0.999 CPUs utilized
     6,129,118,094      instructions:u
         7,518,301      cache-misses:u            #    7.132 % of all cache refs
       105,422,064      cache-references:u        #  104.307 M/sec

And this is with the 32 bit ABI:

        934.800857      task-clock:u (msec)       #    0.999 CPUs utilized
     6,234,900,095      instructions:u
           789,609      cache-misses:u            #    1.130 % of all cache refs
        69,876,685      cache-references:u        #   74.750 M/sec

       0.935685960 seconds time elapsed

The 32 bit ABI is about 8% faster, even though it executes more CPU
instructions.  That is because it wastes less time waiting for the CPU
cache: only 1.1% cache misses instead of 7.1%.

The problem is that the choice of using `int_fast32_t` everywhere has
leaked into the JasPer ABI, and changing it now will break ABI
compatibility.

So for people who can afford an ABI breakage, this patch adds a new
compile-time option `JAS_ENABLE_32BIT` which builds a more efficient
32 bit ABI.