jpc_dec: fix unintended sign extension
By doing `data[0] << 24`, the value is implicitly casted to `signed
int`, and if the highest bit happens to be set, the following cast to
`uint_least32_t` (which has 64 bits on 64 bit CPUs) will cause all
upper 32 bits to be set.  This results in an endless loop (Denial of
Service vulnerability).