Fix incorrect precinct edge detection in RPCL progression order
Why: The condition for edge detection of precincts in jpc_pi_nextrpcl is wrong. (For example the p1_07 codestream used for the conformance contains first the packet for component 1, resolution 0, component 0, position (4,0) before the packet for component 0, resolution 0, component 0, position (8,0), but JasPer reads them in the wrong order.)
How: Fixed the condition according to the JPEG 2000 spec defined as the ISO/IEC 15444-1.