Dequantization using the mid-point reconstruction
Why: In the JPEG 2000 spec, the reconstruction parameter is introduced in the Annex E (Quantization). This reconstruction parameter can be arbitrarily chosen by the decoder, however, it is important to produce the best visual or objective quality for reconstruction. Generally, values for this parameter fall in the range [0, 1) and a common value is 0.5.
How: Added new variable as the reconstruction parameter in jpc_dequantize() and set the value to 0.5.