Fix incorrect checking of the EPH marker
Why: JasPer assumes that there is no successive two bytes out of the range
from 0xff80 to 0xffff in a bitstream. This assumption prevents the
decoder from correct checking the existence of EPH marker (0xff92) when the
third LSB of the Scod parameter is "1".

How: Delete the assumption from the code lines.