jpc_t1{dec,enc}: restore and fix the JPC_PASSTYPE() calls
The `#if 1` code is equivalent, but less clear.  Let's restore the
calls.  This requires fixing the parameter: the `+2` offset is wrong,
but it has been since the very first git commit, so I can't seem to
find out why.

If optimization is really important, we better convert JPC_PASSTYPE()
to an inline function with that formula.  But for now, code clarity
seems more important.