Fix wrong EPH marker checking
If EPH markers are required (by setting the third LSB of Scod parameter to "1" in the COD marker segment, see A.6.1), each packet header in any given tile-part shall be postpended with an EPH marker segment. jpc_dec_lookahead() never returns a value out of the range from 0xff80 to 0xffff, however, a broken bitstream may have a value in this range. Therefore, `if (jpc_dec_lookahead(pkthdrstream) == JPC_MS_EPH)` cannot check the case that the packet header has not been correctly postpended with an EPH marker (= 0xff92).