Hotfix to SOP marker checking
Why: The commit 13390bd said "the `dec->numpkts` holds the number of packets that have been already read. Because the maximum value of `Nsop` is 65535, `dec->numpkts % 65536` is necessary to validate a `Nsop` value in a SOP marker segment.", however, this was wrong.

How: Instead of the `dec->numpkts`, `tile->pi->pktno` is used to validate a `Nsop` because the `Nsop` is defined as packet seqence number in a **tile**. `tile->pi->pktno` holds the packet seqence number in a certain tile.