CMakeLists.txt: remove unimplemented option JAS_MEMORY_LIMIT
This option was added by commit 055568d29d, but there never was an
implementation.  It was never connected to the
`JAS_DEFAULT_MAX_MEM_USAGE` C macro.

In any case, `JAS_DEFAULT_MAX_MEM_USAGE` is flawed, and therefore I
would not like to extend its support, because it is not thread-safe.
This race condition can make libJasPer completely defunct.