jpc_dec: skip clipping if the min/max calculation would overflow
Even though the JPEG2000 standard allows a precision of up to 38 bit
(section A.5.1 table A-11), JasPer can't represent such values In 32
bit mode.  To avoid integer overflows in the min/max value
calculation, which would corrupt the data even more, this commit adds
a check for the precision and omits the clipping stage if the
precision is too large.

While this looks like a hot fix, and we should catch unsupported
precisions earlier, this is also useful for the maximum supported
precision (31 bits) - that calculation (1<<31) would also overflow,
but clipping should be a no-op there.