Fixes Security vulnerability: the MIF decoder opens arbitrary files
The manner in which the MIF codec is enabled has been changed so that no
sane library packager would accidentally enable the MIF codec without
realizing that there are negative security implications.  Now, the
only way to enable the MIF codec is to enable the cmake option called
JAS_DANGEROUS_INTERNAL_TESTING_MODE.  This option defaults to false so
that the MIF codec is not enabled.  Furthermore, this option has multiple
strongly-worded warnings that it should not be enabled except for internal
use by the JasPer development team.