Closes https://github.com/jasper-software/jasper/issues/247
This code change fixes a build problem introduced in 1229a12.  A botched
attempt to fix this problem was made in a0856f1 (which fixed one problem
by creating another).

The code change in this commit moves the point at which the JasPer
OpenGL cmake module is included in the top-level CMakeLists.txt file so
that the main JasPer config header jas_config.h is generated correctly
(with JAS_HAVE_GL_GLUT_H and JAS_HAVE_GLUT_H set properly) and all of
the initialization required for this module is performed prior to its
inclusion.