Reworked the multithreading support.
Purged support for the optional automatic invoking of jas_cleanup
via atexit.

Changed the CMake build process to prefer the use of the POSIX Threads
Library.  This will help to reduce the likelihood of some problems
caused by C11 threading support not interacting well with TSan.

Added --no-lsan and --no-asan options to the build script.  Also, fixed
a typo in a directory name in the same script.