Library: Fixed a serious bug in the memory allocation code (related to jas_realloc). Added the jas_string_tokenize function for splitting a string into tokens. Added the following image-related functions:     jas_image_getfmtbyind     jas_image_getnumfmts     jas_image_setfmtenable Added support for the run-time enabling/disabling of codecs. Added the ability to store multiple file name extensions in each image format table entry. Made the table used to initialize the per-context image-format table private to the library.
Applications:
Changed the -l option of the jasper program to --list-enabled-formats.
Added the --list-all-formats option to the jasper program.
Added the --list-enabled-formats and --list-all-formats options to the
imginfo program.
Added the --enable-format option to the jasper and imginfo programs.
Added the -- (i.e., end-of-options) option to the build script.
Added the test_1 program for performing various simple tests.

Revised the shared-library version information.