Renamed jas_logtype_get_class and jas_logtype_get_priority to jas_logtype_getclass and jas_logtype_getpriority, respectively. Renamed jas_string_tokenize to jas_stringtokenize.
Added support for HEIC encoding of grayscale images.
Fixed a bug in jas_heic_validate.

Added the --enable-all-formats option to the jasper and imginfo programs.
Added the --max-samples and --enable-all-formats options to the jiv program.
Changed the default maximum number of samples for decoding to be unlimited.

Added some code documentation.