Fixes #318.
Eliminated the use of "class" as an identifier, as this will clearly cause
problems if JasPer headers are used in C++ source code.

In order to catch incompatibilities with C++ in the future, a trivial C++
test program has been added.
The CMake build process was adjusted to (optionally) allow the use of a
C++ compiler.
Each use of add_compile_options that sets C compiler options was changed
to use a new function called add_c_compile_options that only sets options
for the C compiler.
Some changes to the CMake module for code sanitizers were made in order to
handle the C++ compiler case.
Several new options were added to the build script.