Fixes #342, #343, #344.
Some command-line options for the jasper, imginfo, imgcmp, and jiv
programs (that are not deemed to be for internal use only) were not
documented in the manual.  Descriptions for these options have now been
added to the manual.

The help information message printed by each of the jasper, imginfo,
imgcmp, and jiv programs has been improved.  A mention of the URL for the
online documentation is now included in all help messages.  A comment
has also been added indicating that not all options are listed in the
help message and that the manual should be checked for a complete list
of options (excluding, of course, the ones that are deemed to be for
internal use only and are deliberately undocumented).